backend/requirements.txt
# FastAPI 核心
fastapi==0.121.2
# ASGI 伺服器（用來運行FastApi）
uvicorn[standard]==0.32.1
pydantic==2.12.4
pydantic-settings==2.12.0

# MongoDB
motor==3.3.2  # 非同步 MongoDB 驅動
pymongo==4.10.1

# 認證與安全
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.6
email-validator==2.3.0

# 工具
python-dotenv==1.0.1


# pip install -r requirements.txt